api-gateway/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
boto3==1.34.14
python-decouple==3.8
